+++ packages/transaction-controller/src/index.ts
@@ -70,7 +70,7 @@ export {
   WalletDevice,
 } from './types';
 export type { EtherscanTransactionMeta } from './utils/etherscan';
-export { ACTIVITY_LOG_HISTORY_PATHS } from './utils/history';
+export { DISPLAYED_TRANSACTION_HISTORY_PATHS } from './utils/history';
 export { determineTransactionType } from './utils/transaction-type';
 export { mergeGasFeeEstimates } from './utils/gas-flow';
 export {